external_data_analysis/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP
textblob>=0.17.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
